move ShellFunctions to stellar-common
